minor testing changes

+++ engine/src/backend/macros.rs
@@ -7,7 +7,7 @@ macro_rules! decode_arithmetic_op {
                         $u.stack.push(Data::INT(l_val $t r_val));
                     },
                     Data::FLOAT(l_val) => {
-                        $u.stack.push(Data::FLOAT(l_val $t r_val as f32));
+                        $u.stack.push(Data::FLOAT(l_val $t r_val as f64));
                     },
                     _ => return InterpretResult::COMPILE_ERROR,
                 }
@@ -15,7 +15,7 @@ macro_rules! decode_arithmetic_op {
             Data::FLOAT(r_val) => {
                 match $u.stack.pop() {
                     Data::INT(l_val) => {
-                        $u.stack.push(Data::FLOAT(l_val as f32 $t r_val));
+                        $u.stack.push(Data::FLOAT(l_val as f64 $t r_val));
                     },
                     Data::FLOAT(l_val) => {
                         $u.stack.push(Data::FLOAT(l_val $t r_val));
@@ -37,7 +37,7 @@ macro_rules! decode_comparison_op {
                         $u.stack.push(Data::BOOL(l_val $t r_val))
                     },
                     Data::FLOAT(l_val) => {
-                        $u.stack.push(Data::BOOL(l_val $t r_val as f32))
+                        $u.stack.push(Data::BOOL(l_val $t r_val as f64))
                     },
                     _ => return InterpretResult::COMPILE_ERROR
                 }
@@ -45,7 +45,7 @@ macro_rules! decode_comparison_op {
             Data::FLOAT(r_val) => {
                 match $u.stack.pop() {
                     Data::INT(l_val) => {
-                        $u.stack.push(Data::BOOL((l_val as f32) $t r_val))
+                        $u.stack.push(Data::BOOL((l_val as f64) $t r_val))
                     },
                     Data::FLOAT(l_val) => {
                         $u.stack.push(Data::BOOL(l_val $t r_val))
@@ -67,7 +67,7 @@ macro_rules! decode_equality_op {
                         $u.stack.push(Data::BOOL(l_val $t r_val))
                     },
                     Data::FLOAT(l_val) => {
-                        $u.stack.push(Data::BOOL(l_val $t r_val as f32))
+                        $u.stack.push(Data::BOOL(l_val $t r_val as f64))
                     },
                     _ => return InterpretResult::COMPILE_ERROR
                 }
@@ -75,7 +75,7 @@ macro_rules! decode_equality_op {
             Data::FLOAT(r_val) => {
                 match $u.stack.pop() {
                     Data::INT(l_val) => {
-                        $u.stack.push(Data::BOOL(l_val as f32 $t r_val))
+                        $u.stack.push(Data::BOOL(l_val as f64 $t r_val))
                     },
                     Data::FLOAT(l_val) => {
                         $u.stack.push(Data::BOOL(l_val $t r_val))
